feat: handle image absence
