+++ react-ts/src/components/detailed-card-image/ImageNotFound.tsx
@@ -1,0 +1,10 @@
+import './ImageNotFound.css';
+const NotFoundImage = () => {
+  return (
+    <div className="not-found detailed-card__image-wrapper">
+      No image available
+    </div>
+  );
+};
+
+export default NotFoundImage;
